:recycle: Move all cells to individual files
